dev-cit-3.0-windows_support: modified key generation mechanism in bitlocker setup script, changed long to uint in service src file, changed location of vRTM in PA properties file, and added code to stop and delete unlock service while uninstall

+++ bitlocker_service/BitLocker/BitLocker.cs
@@ -23,13 +23,13 @@ public enum ServiceState
 [StructLayout(LayoutKind.Sequential)]
 public struct ServiceStatus
 {
-    public long dwServiceType;
+    public uint dwServiceType;
     public ServiceState dwCurrentState;
-    public long dwControlsAccepted;
-    public long dwWin32ExitCode;
-    public long dwServiceSpecificExitCode;
-    public long dwCheckPoint;
-    public long dwWaitHint;
+    public uint dwControlsAccepted;
+    public uint dwWin32ExitCode;
+    public uint dwServiceSpecificExitCode;
+    public uint dwCheckPoint;
+    public uint dwWaitHint;
 };
 
 namespace BitLocker
